+++ src/domain/community/preferences/user.preference.definition.interface.ts
